refactor: split out `handleWebSearchToolCall` method
